chore(skeleton): move to universal core
