+++ src/components/client/core/index.ts
@@ -4,7 +4,6 @@ export { default as Drawer } from "./Drawer/Drawer";
 export { default as Icon } from "./Icon/Icon";
 export { default as Menu } from "./Menu/Menu";
 export { default as Modal } from "./Modal/Modal";
-export { default as Skeleton } from "./Skeleton/Skeleton";
 export { default as Slider } from "./Slider/Slider";
 export { default as Tabs } from "./Tabs/Tabs";
 export { default as Toggle } from "./Toggle/Toggle";
@@ -22,7 +21,6 @@ export type {
   MenuItemGroupRecord,
 } from "./Menu/Menu";
 export type { Props as ModalProps } from "./Modal/Modal";
-export type { Props as SkeletonProps } from "./Skeleton/Skeleton";
 export type { Props as SliderProps, SliderMarkerRecord } from "./Slider/Slider";
 export type { Props as TabsProps, TabRecord } from "./Tabs/Tabs";
 export type { Props as ToggleProps } from "./Toggle/Toggle";
