+++ src/components/universal/core/Skeleton/Skeleton.recipe.ts
