+++ src/components/universal/core/index.ts
@@ -1,6 +1,7 @@
 export { default as Badge } from "./Badge/Badge";
 export { default as Card, CardHeader, CardBody, CardFooter } from "./Card/Card";
 export { default as Image } from "./Image/Image";
+export { default as Skeleton } from "./Skeleton/Skeleton";
 export { default as Spinner } from "./Spinner/Spinner";
 export { default as Text } from "./Text/Text";
 
@@ -11,5 +12,6 @@ export type {
   CardFooterProps,
 } from "./Card/Card";
 export type { Props as ImageProps } from "./Image/Image";
+export type { Props as SkeletonProps } from "./Skeleton/Skeleton";
 export type { Props as SpinnerProps } from "./Spinner/Spinner";
 export type { Props as TextProps } from "./Text/Text";
